+++ examples/gulp-requirejs/with-global-component/src/components/p-demo/p-demo.js
@@ -1,5 +1,0 @@
-'use strict';
-require(['rDemo'], function(rDemo){
-    rDemo.init();
-
-});