+++ history.md
@@ -4,6 +4,7 @@
 * [ADD] `gulp-requirejs`, `gulp-rollup` 新增 `--remote` 选项
 * [ADD] 新增 `yyl make` 功能
 * [EDIT] `yyl update` 功能完成
+* [ADD] `gulp-requirejs`, `gulp-rollup` 支持在 `p-xx.jade` 里面直接引用 `p-xx.scss` 构建工具会自动转义成 `xx.css`
 
 ## 2.15.23 (2017-12-01)
 * [FIX] 修复 `gulp-requirejs`, `gulp-rollup` 运行 会报错问题